Move coder struct to its own file
To ease development and find stuff
